modulo-produtos/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
alembic==1.16.4
pydantic==2.11.7
pydantic-settings==2.3.4
python-multipart==0.0.9
 Pillow==10.4.0

